// src/frontend-applications/agent-portal/modules/system-log/model/LogFileProperty.ts
/**
 * Copyright (C) 2006-2022 c.a.p.e. IT GmbH, https://www.cape-it.de
 * --
 * This software comes with ABSOLUTELY NO WARRANTY. For details, see
 * the enclosed file LICENSE for license information (GPL3). If you
 * did not receive this file, see https://www.gnu.org/licenses/gpl-3.0.txt.
 * --
 */

export enum LogFileProperty {

    ACCESS_TIME = 'AccessTime',

    ACCESS_TIME_UNIX = 'AccessTimeUnix',

    CREATE_TIME = 'CreateTime',

    CREATE_TIME_UNIX = 'CreateTimeUnix',

    DISPLAY_NAME = 'DisplayName',

    FILE_NAME = 'Filename',

    FILE_SIZE = 'Filesize',

    FILE_SIZE_RAW = 'FilesizeRaw',

    ID = 'ID',

    MODIFIY_TIME = 'ModifyTime',

    MODIFY_TIME_UNIX = 'ModifyTimeUnix',

    CONTENT = 'Content',

}